% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_data_with_reliability}
\alias{simulate_data_with_reliability}
\title{Simulate dataset with true scores, errors and total scores}
\usage{
simulate_data_with_reliability(n, rxx, m = 100, sd = 15)
}
\arguments{
\item{n}{number of rows}

\item{rxx}{reliability of the test}

\item{m}{mean}

\item{sd}{standard deviation}
}
\value{
tibble with columns for true scores, observed scores and errors
}
\description{
Simulate dataset with true scores, errors and total scores
}
\examples{
simulate_data_with_reliability(20, .8)
}
